move simulation.py to striatum
