+++ simulation/simulation_exp4p.py
@@ -1,7 +1,7 @@
 from striatum.storage import history
 from striatum.storage import model
 from striatum.bandit import exp4p
-import simulation as sm
+from striatum import simulation as sm
 import numpy as np
