+++ striatum/simulation.py
